Significant redesign of machine core to animate multiple properties
Previously we had to select one property to animate at a time and could
only set static values on others.

Now we can animate anywhere between none and all of the properties at
the same time. This also makes it easier to add new properties so we
can animate hue and lightness in the future.