Simplify render by making properties either animated or constant
Rather than having to inspect divider and use constant or value at render
time we just set value equal to constant when divider is zero.

The only down side is that we cant manually "spin" with constant while
animating but an additional modal rotation feature could be added if we
wanted this live control.